+++ initial-data-load/src/database/mongoDb/constants.ts
@@ -1,0 +1,2 @@
+export const ATTRIBUTES_COLLECTION_NAME = 'attributes';
+export const POINT_ATTRIBUTES_COLLECTION_NAME = 'pointAttributes';