+++ postgis_init/entrypoint_initdb/99_20220513000001_create_layers_tables.sh
@@ -19,15 +19,16 @@ psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" <<-E
 
   CREATE TABLE map_layers (
     reference_id           VARCHAR         NOT NULL,
-    geo_reference_id       VARCHAR         NOT NULL references layer_geo_data(reference_id), 
+    geo_reference_id       VARCHAR         references layer_geo_data(reference_id), 
     layer_type             VARCHAR         NOT NULL, 
     category               VARCHAR         NOT NULL, 
     title                  VARCHAR         NOT NULL,
-    attribute_id           VARCHAR         NOT NULL references attributes(attribute_id),
-    attribute_description  JSONB           NOT NULL,
-    styles                 JSONB           NOT NULL,
-    legend                 JSONB           NOT NULL,
+    attribute_id           VARCHAR         references attributes(attribute_id),
+    attribute_description  JSONB,
+    styles                 JSONB,
+    legend                 JSONB[],
     layer_options          JSONB           NOT NULL,
+    layers                 JSONB[],
 
     created_at TIMESTAMP WITH TIME ZONE NULL DEFAULT NOW(),
     updated_at TIMESTAMP WITH TIME ZONE NULL DEFAULT NOW(),